add contact - added keyboard

+++ app/src/main/java/edu/ib/visionhelper/call/CallListElement.kt
@@ -1,6 +1,6 @@
 package edu.ib.visionhelper.call
 
-class CallListElement(var contactName: String, var contactNumber: Int) {
+class CallListElement(var contactName: String, var contactNumber: Long) {
 
     override fun toString(): String {
         return "CallListElement(contactName='$contactName', contactNumber=$contactNumber)"
